fix data type for run_time

+++ db/migrations/20210218124156-create-test-meta-data.js
@@ -39,7 +39,7 @@ module.exports = {
         allowNull: false,
       },
       run_time: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL(10, 3),
         allowNull: false,
       },
       message: {
